+++ pkg/storage/storagewrappers/request_test.go
@@ -28,14 +28,17 @@ func TestRequestStorageWrapper(t *testing.T) {
 			tuple.NewTupleKey("doc:1", "viewer", "user:maria"),
 		}
 
-		br := NewRequestStorageWrapper(mockDatastore, requestContextualTuples, maxConcurrentReads,
+		br := NewRequestStorageWrapperWithCache(mockDatastore, requestContextualTuples, maxConcurrentReads,
 			&shared.SharedDatastoreResources{
 				CheckCache: mockCache,
 				Logger:     logger.NewNoopLogger(),
 			}, config.CacheSettings{
 				CheckIteratorCacheEnabled: true,
 				CheckCacheLimit:           1,
-			}, logger.NewNoopLogger())
+			},
+			logger.NewNoopLogger(),
+			Check,
+		)
 		require.NotNil(t, br)
 
 		// assert on the chain
@@ -66,7 +69,15 @@ func TestRequestStorageWrapper(t *testing.T) {
 			tuple.NewTupleKey("doc:1", "viewer", "user:maria"),
 		}
 
-		br := NewRequestStorageWrapper(mockDatastore, requestContextualTuples, maxConcurrentReads, &shared.SharedDatastoreResources{Logger: logger.NewNoopLogger()}, config.CacheSettings{}, logger.NewNoopLogger())
+		br := NewRequestStorageWrapperWithCache(
+			mockDatastore,
+			requestContextualTuples,
+			maxConcurrentReads,
+			&shared.SharedDatastoreResources{Logger: logger.NewNoopLogger()},
+			config.CacheSettings{},
+			logger.NewNoopLogger(),
+			Check,
+		)
 		require.NotNil(t, br)
 
 		// assert on the chain
@@ -90,7 +101,7 @@ func TestRequestStorageWrapper(t *testing.T) {
 			tuple.NewTupleKey("doc:1", "viewer", "user:maria"),
 		}
 
-		br := NewRequestStorageWrapperForListAPIs(mockDatastore, requestContextualTuples, maxConcurrentReads)
+		br := NewRequestStorageWrapper(mockDatastore, requestContextualTuples, maxConcurrentReads)
 		require.NotNil(t, br)
 
 		// assert on the chain